Separate On Demand Sync into a VFService
