+++ com.vrcfury.vrcfury/Editor/VF/Feature/OnDemandSyncBuilder.cs
@@ -8,94 +8,21 @@ using VF.Inspector;
 using VF.Model.Feature;
 using VF.Service;
 using VF.Utils.Controller;
+using VRC.SDK3.Avatars.ScriptableObjects;
 
 namespace Editor.VF.Feature {
     public class OnDemandSyncBuilder : FeatureBuilder<OnDemandSync> {
-        [VFAutowired] private readonly MathService math;
-        [VFAutowired] private readonly DirectBlendTreeService directTree;
-        [VFAutowired] private readonly FrameTimeService frameTimeService;
-        [VFAutowired] private readonly SmoothingService smoothingService;
-        
-        [FeatureBuilderAction()]
+        [VFAutowired] private readonly OnDemandSyncService onDemandSyncService;
+
+        [FeatureBuilderAction]
         public void Apply() {
-            var fx = GetFx();
-            var layers = fx.GetRaw().GetLayers();
             foreach (var param in model.parameters) {
-                manager.GetParams().GetParam(param).networkSynced = false;
-            }
-            var pointer = fx.NewInt("SyncPointer", synced: true);
-            var localPointer = fx.NewInt("LocalPointer"); // Use a local only pointer to avoid race condition
-            var data = fx.NewFloat("SyncData", synced: true);
-            var sending = fx.NewFloat("IsSending");
-
-            var txLayer = fx.NewLayer("Send");
-            var txIdle = txLayer.NewState("Idle");
-            txIdle.Drives(sending, 0, true);
-            var txAdder = txLayer.NewState("Adder");
-            var txResetter = txLayer.NewState("Reset");
-            txResetter.Drives(localPointer, 0, local: true)
-                .TransitionsToExit()
-                .When(fx.Always());
-            txAdder.TransitionsTo(txResetter)
-                .When(localPointer.IsGreaterThan(model.parameters.Count - 1));
-            txAdder.DrivesDelta(localPointer, 1)
-                .TransitionsToExit()
-                .When(fx.Always());
-            
-            for (int i = 0; i < model.parameters.Count; i++) {
-                var src = new VFAFloat(fx.GetRaw().GetParam(model.parameters[i]));
-                var lastSrc = fx.NewFloat($"last{src.Name()}");
-                directTree.Add(math.MakeCopier(src, lastSrc)); 
-                var srcDiff = math.Subtract(src, lastSrc);
-                var maintain = math.MakeMaintainer(srcDiff); // Maintain while in send animation
-                directTree.Add(sending, maintain);
-                
-                var sendInstantState = txLayer.NewState($"Send Instant {model.parameters[i]}");
-                sendInstantState
-                    .DrivesCopy(data, src)
-                    .Drives(pointer, i)
-                    .Drives(sending, 1, true)
-                    .TransitionsToExit()
-                    .WithTransitionExitTime(0.1f)
-                    .When(fx.Always());
-                txIdle.TransitionsTo(sendInstantState)
-                    .When(fx.IsLocal().IsTrue().And(srcDiff.IsGreaterThan(0).Or(srcDiff.IsLessThan(0))));
-            }
-
-            for (int i = 0; i < model.parameters.Count; i++) {
-                var src = new VFAFloat(fx.GetRaw().GetParam(model.parameters[i]));
-                var sendState = txLayer.NewState($"Send {model.parameters[i]}");
-                sendState
-                    .DrivesCopy(data, src)
-                    .DrivesCopy(pointer, localPointer)
-                    .Drives(sending, 1, true)
-                    .TransitionsTo(txAdder)
-                    .WithTransitionExitTime(0.1f)
-                    .When(fx.Always());
-                txIdle.TransitionsTo(sendState)
-                    .When(fx.IsLocal().IsTrue().And(localPointer.IsEqualTo(i)));
-            }
-            
-            var rxLayer = fx.NewLayer("Receive");
-            var rxIdle = rxLayer.NewState("Idle");
-            var smoothedDict = new Dictionary<string, string>();
-            for (int i = 0; i < model.parameters.Count; i++) {
-                var dst = new VFAFloat(fx.GetRaw().GetParam(model.parameters[i]));
-                var rxState = rxLayer.NewState($"Receive {model.parameters[i]}");
-                rxState.DrivesCopy(dst, data, false)
-                    .TransitionsToExit()
-                    .When(fx.Always());
-                rxIdle.TransitionsTo(rxState)
-                    .When(pointer.IsEqualTo(i).And(fx.IsLocal().IsFalse()));
-                
-                var dstSmoothed = smoothingService.Smooth($"{dst.Name()}/Smoothed", dst, 0.1f, false);
-                smoothedDict[dst.Name()] = dstSmoothed.Name();
-                fx.GetRaw().RewriteParameters(name => {
-                    if (smoothedDict.TryGetValue(name, out var smoothed)) {
-                        return smoothed;
-                    }
-                    return name;
-                }, false, layers.Select(l => l.stateMachine).ToArray());
+                var vrcParam = manager.GetParams().GetParam(param);
+                if(vrcParam == null) continue;
+                if(vrcParam.valueType != VRCExpressionParameters.ValueType.Float) continue;
+                var controllerParam = GetFx().GetRaw().GetParam(param);
+                if(controllerParam == null) continue;
+                onDemandSyncService.SetParameterOnDemandSync(new VFAFloat(controllerParam));
             }
         }
